tts-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
piper-tts>=1.2.0
numpy>=1.26.0
librosa>=0.10.0
soundfile>=0.12.0
pydantic-settings>=2.2.1
kokoro>=0.9.2
